drm: Convert into a CogShell implementation
Turn the cog-platform-drm.c plug-in into a CogShell implementation,
with support for multiple views, only one of them show at a time.

This is a first pass that gets graphics output working, but no input
handling is done. Handling of lower level DRM/KMS is left to to a new
"pdrm" platform utility library which abstracts away most of the
complexity of using libdrm and libgbm.